COMPARISON regex - from higher to lower specificity
Hi I noticed the regex wasn't matching when I tried it out in a regex tester, then I noticed it would match a string of ">" but fail to match ">=". Looked at the pattern and realised the ">" was matching first. Leex regex implementation may work some other way but I'm giving the PR anyway, just in case.